% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{c_chart}
\alias{c_chart}
\title{C chart}
\usage{
c_chart(data, x)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}
}
\value{
An object of class c_chart
}
\description{
Generate a C control chart
}
\examples{
data(test_c)
c_chart(test_c, defect)
}

