% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{c_chart_}
\alias{c_chart_}
\title{C chart}
\usage{
c_chart_(data, x)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}
}
\value{
An object of class c_chart
}
\description{
Generate a C control chart object
}

