% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_range}
\alias{diff_range}
\title{Computes  max(x) - min(x)}
\usage{
diff_range(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{other numeric vectors}
}
\value{
A numeric vector of length one
}
\description{
Computes  max(x) - min(x)
}
\examples{
diff_range(x = 1:5)
diff_range(x = c(1:5, NA), na.rm = TRUE)
}

