% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{i_chart}
\alias{i_chart}
\title{I chart}
\usage{
i_chart(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{step of the moving range}
}
\value{
An object of class i_chart
}
\description{
Generate a I control chart
}
\examples{
data(test_i_mr)
i_chart(test_i_mr, cost, 2)
}

