% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{i_chart_}
\alias{i_chart_}
\title{I chart}
\usage{
i_chart_(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{numeric integer step of the moving range}
}
\value{
An object of class i_chart
}
\description{
Generate a I control chart
}

