% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{mr_chart}
\alias{mr_chart}
\title{MR chart}
\usage{
mr_chart(data, x, mr_step)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{mr_step}{step of the moving range}
}
\value{
An object of class mr_chart
}
\description{
Generate an MR control chart
}
\examples{
# Example with moving range of 2 (minumum moving range window size)
data(test_i_mr)
mr_chart(test_i_mr, cost, mr_step = 2)

}

