% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{mr_chart_}
\alias{mr_chart_}
\title{MR chart}
\usage{
mr_chart_(data, x, mr_step)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{mr_step}{step of the moving range}
}
\value{
An object of class mr_chart
}
\description{
Generate an MR control chart
}
\examples{

}

