% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{np_chart}
\alias{np_chart}
\title{NP chart}
\usage{
np_chart(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{size of each sample}
}
\value{
An object of class np_chart
}
\description{
Generate a NP control chart
}
\examples{
data(test_np)
np_chart(test_np, non_conforming_cans, sample_size)
}

