% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{np_chart_}
\alias{np_chart_}
\title{NP chart}
\usage{
np_chart_(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{size of each sample}
}
\value{
An object of class np_chart
}
\description{
Generate a NP control chart
}

