% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{p_chart_}
\alias{p_chart_}
\title{P chart}
\usage{
p_chart_(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{size of each sample}
}
\value{
An object of class p_chart
}
\description{
Generate a P control chart
}

