% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{r}
\alias{r}
\title{Generate a R object}
\usage{
r(data, x, g, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{sigma}{standard deviation of variable x (if known)}
}
\value{
An object of class xbar-R
}
\description{
Generate a R object
}
\examples{
data(ingots)
r(ingots, kg, group)
r(ingots, kg, group, sigma=0.05)
}

