% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{r_}
\alias{r_}
\title{SE version of r}
\usage{
r_(data, x, g, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{sigma}{standard deviation of variable x (if known)}
}
\value{
An object of class r
}
\description{
SE version of r
}

