% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{s}
\alias{s}
\title{Generate a S object}
\usage{
s(data, x, g, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{sigma}{standard deviation of x variable (if known)}
}
\value{
An object of class s
}
\description{
Generate a S object
}
\examples{
data(ingots)
s(production, x = dim_mm, g = batch)
s(production, x = dim_mm, g = batch, sigma=2)
}

