% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_center}
\alias{set_center}
\title{set_center}
\usage{
set_center(data, new_center)
}
\arguments{
\item{data}{an object of class spc}

\item{new_center}{a single value (constant center) or a numeric vector}
}
\value{
An object of class spc
}
\description{
Set center parameters in control chart object
}

