% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_constant}
\alias{set_constant}
\title{Utility function. Returns arg_one if k is NULL otherwise returns arg_two}
\usage{
set_constant(arg_one, arg_two, k = NULL)
}
\arguments{
\item{k}{control chart parameter (ex: mu or sigma) set by the user or NULL}
}
\description{
Utility function. Returns arg_one if k is NULL otherwise returns arg_two
}

