% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_function}
\alias{set_function}
\title{Utility factory function. Returns fun if k is null, else returns f(x){k}}
\usage{
set_function(fun, k = NULL)
}
\arguments{
\item{fun}{stat_fun or shape_fun}

\item{k}{control chart parameter (ex: mu or sigma) set by the user or NULL}
}
\description{
Utility factory function. Returns fun if k is null, else returns f(x){k}
}

