% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_lcl}
\alias{set_lcl}
\title{set_lcl}
\usage{
set_lcl(data, new_lcl)
}
\arguments{
\item{data}{an object of class spc}

\item{new_lcl}{a single value (constant lcl) or a numeric vector}
}
\value{
An object of class spc
}
\description{
Set lcl parameters in control chart object
}

