% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_param.c_chart}
\alias{set_param.c_chart}
\title{set_param.c_chart}
\usage{
\method{set_param}{c_chart}(data, param_c)
}
\arguments{
\item{data}{an object of class c_chart}

\item{param_c}{the C parameter of the c chart}
}
\value{
An object of class c_chart
}
\description{
Set the Poisson's distribution parameter C in the c chart.
}

