% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_param.p_chart}
\alias{set_param.p_chart}
\title{set_param.p_chart}
\usage{
\method{set_param}{p_chart}(data, param_p)
}
\arguments{
\item{data}{an object of class p_chart}

\item{param_p}{the P parameter of the p chart}
}
\value{
An object of class p_chart
}
\description{
If the fraction of non-conforming items P is known, then you can set it in the chart.
}

