% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_user_defined_limits.R
\name{set_param}
\alias{set_param}
\title{set_param}
\usage{
set_param(data, ...)
}
\arguments{
\item{data}{a SPC object}
}
\description{
Generic S3 method for setting chart parameters if known.
}

