% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_param.s}
\alias{set_param.s}
\title{set_param.s}
\usage{
\method{set_param}{s}(data, param_sigma)
}
\arguments{
\item{data}{an object of class s}

\item{param_sigma}{the parameter of the s chart}
}
\value{
An object of class s
}
\description{
Set the s chart parameter sigma.
}

