% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_user_defined_limits.R
\name{set_param_generic_}
\alias{set_param_generic_}
\title{set_param_generic}
\usage{
set_param_generic_(data, param, value)
}
\arguments{
\item{data}{an object of class spc}

\item{param}{parameter to be replaced (center, lcl, ucl or stat)}

\item{value}{a single numeric value (constant)}
}
\value{
An object of class spc
}
\description{
Set user defined parameters in control chart object. Example: set user defined
UCL, LCL, center or even stats.
}

