% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_ucl}
\alias{set_ucl}
\title{set_ucl}
\usage{
set_ucl(data, new_ucl)
}
\arguments{
\item{data}{an object of class spc}

\item{new_ucl}{a single value (constant ucl) or a numeric vector}
}
\value{
An object of class spc
}
\description{
Set ucl parameters in control chart object
}

