% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{u_chart}
\alias{u_chart}
\title{U chart}
\usage{
u_chart(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{number of inspection units per sample (sample size within each inspection unit is assumed constant)}
}
\value{
An object of class u_chart
}
\description{
Generate a U control chart
}
\examples{
data(test_u)
#u_chart(test_u, non_conformities, number_of_inspection_units)
}

