% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{u_chart_}
\alias{u_chart_}
\title{U chart}
\usage{
u_chart_(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{number of inspection units per sample (sample size within each inspection unit is assumed constant)}
}
\value{
An object of class u_chart
}
\description{
Generate a U control chart
}

