% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_limits_functions.R
\name{ucl_fun.xbar_r}
\alias{ucl_fun.xbar_r}
\title{ucl_fun xbar-r}
\usage{
\method{ucl_fun}{xbar_r}(value = NULL, ...)
}
\arguments{
\item{value}{ucl value set by the user}

\item{class}{class of the chart}
}
\description{
This function returns the formula for calculating the lcl for the xbar-r
chart if no value is provided by the user. If a value is provided, then
the function returns value.
}
\examples{
ucl_fun("xbar_r", 2)
ucl_fun("xbar_r")
}

