% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{xbar_r}
\alias{xbar_r}
\title{Generate a xbar R object}
\usage{
xbar_r(data, x, g, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{mu}{average value of x variable (if known)}

\item{sigma}{standard deviation of x variable (if known)}
}
\value{
An object of class xbar-R
}
\description{
Generate a xbar R object
}
\examples{
data(ingots)
xbar_r (ingots, kg, group)
xbar_r(ingots, kg, group, mu=1, sigma=0.01)
}

