% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{c_chart_}
\alias{c_chart_}
\title{C chart}
\usage{
c_chart_(data, x, g, center = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{This variable is not used in the c chart. It is set to be equal to x by default in the NSE function.}

\item{center}{parameter of the c chart, if known}
}
\value{
An object of class c_chart
}
\description{
Generate a C control chart object
}

