% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{p_chart}
\alias{p_chart}
\title{P chart}
\usage{
p_chart(data, x, g, p = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{size of each sample}

\item{p}{parameter of the p chart}
}
\value{
An object of class p_chart
}
\description{
Generate a P control chart
}
\examples{
data(test_p)
p_chart(test_p, defect, sample_size)
}

