% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.spc}
\alias{plot.spc}
\title{plot.spc
plot an object of class spc}
\usage{
\method{plot}{spc}(data, colour = c("darkblue", "darkred", "darkred",
  "darkred"))
}
\arguments{
\item{data}{spc object}

\item{colour}{colours for statistics, center line, ucl and lcl}
}
\value{
A ggplot object
}
\description{
plot.spc
plot an object of class spc
}
\examples{
#Ingots Data
ingots_xbar_r <- xbar_r (ingots, kg, group)
plot(ingots_xbar_r)
#Pistonrings Data
pistonrings_xbar_r <- xbar_r (pistonrings[1:125,], diameter, sample)
plot(pistonrings_xbar_r)
}

