% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{s_}
\alias{s_}
\title{SE version of s}
\usage{
s_(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}
}
\value{
An object of class r
}
\description{
SE version of s
}

