% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_function}
\alias{set_function}
\title{Utility factory function. Returns fun if k is null, else returns f(x){k}}
\usage{
set_function(fun, k = NULL)
}
\arguments{
\item{fun}{stat_fun or shape_fun}

\item{k}{control chart parameter (ex: mu or sigma) set by the user or NULL}

\item{n}{sample size}

\item{const}{control chart constant, if required}
}
\description{
Utility factory function. Returns fun if k is null, else returns f(x){k}
}
\examples{
set_function(mean)         # Returns mean
set_function(mean, k = 10) # Returns function(x, ...){ 10 }
}

