% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{set_param_generic}
\alias{set_param_generic}
\title{set_param_generic}
\usage{
set_param_generic(data, param, value)
}
\arguments{
\item{data}{an object of class spc}

\item{param}{parameter to be replaced (center, lcl, ucl or stat)}

\item{value}{a single numeric value (constant)}
}
\value{
An object of class spc
}
\description{
Set user defined parameters in control chart (SPC) object. For instance: set user defined
UCL, LCL, center or even stats.
}
\examples{
 data("test_p")
 p <- p_chart(test_p,defect,sample_size)
 set_param_generic(p,10,lcl)
}

