% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_set_param.R
\name{set_param_generic_}
\alias{set_param_generic_}
\title{set_param_generic}
\usage{
set_param_generic_(data, new_param, replace_param)
}
\arguments{
\item{data}{an object of class spc}

\item{new_param}{a single value (constant) or a numeric vector}

\item{replace_param}{parameter to be replaced (center, lcl or ucl)}
}
\value{
An object of class spc
}
\description{
Set user defined parameters in control chart object
}

