% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{xbar_r_}
\alias{xbar_r_}
\title{xbar_r_
SE version of xbar-R}
\usage{
xbar_r_(data, x, g, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{mu}{average value of x variable (if known)}

\item{sigma}{standard deviation of x variable (if known)}
}
\value{
An object of class xbar-r
}
\description{
xbar_r_
SE version of xbar-R
}

