% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{xbar_s}
\alias{xbar_s}
\title{Generate a xbar-s object}
\usage{
xbar_s(data, x, g)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}
}
\value{
An object of class xbar-R
}
\description{
Generate a xbar-s object
}
\examples{
data(ingots)
xbar_s (production, x = dim_mm, g = batch)
}

