% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{xbar_s_}
\alias{xbar_s_}
\title{SE version of xbar-S}
\usage{
xbar_s_(data, x, g, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{variable to be used}

\item{g}{group variable}

\item{mu}{average value of x variable (if known)}

\item{sigma}{standard deviation of x variable (if known)}
}
\value{
An object of class xbar-s
}
\description{
SE version of xbar-S
}

